backend/requirements.txt
# aiofiles==23.1.0
# celery==5.2.7
# fastapi==0.95.0
# flower==1.2.0
# Jinja2==3.1.2
# pytest==7.2.2
# redis==4.5.4
# requests==2.28.2
# uvicorn==0.21.1
# httpx==0.23.3
aiofiles==24.1.0
alembic==1.16.4
asyncpg==0.30.0
celery==5.5.3
fastapi==0.116.1
flower==1.2.0
httpx==0.28.1
Jinja2==3.1.6
# pydantic==1.10.22
pydantic_settings==2.10.1
pytest==8.4.1
redis==6.4.0
requests==2.32.4
uvicorn==0.35.0
psycopg-binary==3.2.9
pyjwt==2.10.1
bcrypt==4.3.0
passlib==1.7.4